fix some windows compile errors

+++ src/igor/Core/ExtractFeatures.cpp
@@ -143,13 +143,13 @@ int ExtractFeatures::getVAnchor4Seq(string seq_str, Alignment_data v_alig)
     cdr3_v_gene_anch = UMap_v_CDR3_anchors[v_alig.gene_name];
     int v_ins_correction = std::count_if(v_alig.get_all_insertions().begin(), v_alig.get_all_insertions().end(),
                                          // Lambda function for condition
-                                         [cdr3_v_gene_anch](int inss) { return (inss <= cdr3_v_gene_anch); });
+                                         [cdr3_v_gene_anch](size_t inss) { return (inss <= cdr3_v_gene_anch); });
     cdr3_v_gene_anch = cdr3_v_gene_anch + v_ins_correction;
 
     int cdr3_v_read_anch = cdr3_v_gene_anch + v_alig.offset;
     int dels_correction = std::count_if(v_alig.get_all_deletions().begin(), v_alig.get_all_deletions().end(),
                                         // Lambda function for condition
-                                        [cdr3_v_read_anch](int dels) { return (dels <= cdr3_v_read_anch); });
+                                        [cdr3_v_read_anch](size_t dels) { return (dels <= cdr3_v_read_anch); });
     cdr3_v_read_anch = cdr3_v_read_anch + dels_correction; // ins_size before the cdr3 a
 
     return cdr3_v_read_anch;
@@ -180,13 +180,13 @@ int ExtractFeatures::getJAnchor4Seq(string seq_str, Alignment_data j_alig)
     cdr3_j_gene_anch = UMap_j_CDR3_anchors[j_alig.gene_name];
     int j_ins_correction = std::count_if(j_alig.get_all_insertions().begin(), j_alig.get_all_insertions().end(),
                                          // Lambda function for condition
-                                         [cdr3_j_gene_anch](int inss) { return (inss <= cdr3_j_gene_anch); });
+                                         [cdr3_j_gene_anch](size_t inss) { return (inss <= cdr3_j_gene_anch); });
     cdr3_j_gene_anch = cdr3_j_gene_anch + j_ins_correction;
 
     int cdr3_j_read_anch = cdr3_j_gene_anch + j_alig.offset;
     int j_dels_correction = std::count_if(j_alig.get_all_deletions().begin(), j_alig.get_all_deletions().end(),
                                           // Lambda function for condition
-                                          [cdr3_j_read_anch](int dels) { return (dels <= cdr3_j_read_anch); });
+                                          [cdr3_j_read_anch](size_t dels) { return (dels <= cdr3_j_read_anch); });
 
     // In order to get the phelanine or triptophan in the sequence +3.
     cdr3_j_read_anch = cdr3_j_read_anch + j_dels_correction + 3; // ins_size before the cdr3 anchor
