+++ src/igor/Core/Hypermutationglobalerrorrate.cpp
@@ -156,7 +156,7 @@ Hypermutation_global_errorrate::Hypermutation_global_errorrate(size_t nmer_width
         learn_on_j = false;
 
     //Initialize adressing vector
-    for (int ii = (mutation_Nmer_size - 1); ii != -1; --ii) {
+    for (size_t ii = (mutation_Nmer_size - 1); ii != -1; --ii) {
         adressing_vector.emplace_back(pow(4, ii));
     }
